chore: move entity tabs
