+++ apps/ai-dial-admin/src/components/ApplicationRunners/View/TabsContent.tsx
@@ -6,7 +6,7 @@ import AppRunnerApplications from '@/src/components/ApplicationRunners/Configura
 import AppRunnerFeatures from '@/src/components/ApplicationRunners/ConfigurationView/Features';
 import ApplicationParametersTab from '@/src/components/Applications/ParametersTab/ParametersTab';
 import EntityRoutes from '@/src/components/EntityView/AppRoute/AppRoute';
-import EntityAudit from '@/src/components/EntityView/Audit/EntityAudit';
+import EntityAudit from '@/src/components/EntityTabs/Audit/EntityAudit';
 import EntityInterceptors from '@/src/components/EntityView/Interceptors/Interceptors';
 import { DialApplicationScheme } from '@/src/models/dial/application';
 import { DialInterceptor } from '@/src/models/dial/interceptor';
