+++ apps/ai-dial-admin/src/components/Assets/Toolsets/View/View.tsx
@@ -9,7 +9,7 @@ import { getToolsets, moveToolsets, removeToolset, updateToolset } from '@/src/a
 import { addNewVersion, getEntityForUpdate, getIsNeedToMove } from '@/src/components/Assets/utils';
 import AssetHeader from '@/src/components/EntityHeaderControls/AssetHeader';
 import { JsonConfiguration } from '@/src/components/EntityHeaderControls/models';
-import EntityJsonEditor from '@/src/components/EntityView/JsonEditor/JsonEditor';
+import EntityJsonEditor from '@/src/components/EntityTabs/JsonEditor/JsonEditor';
 import AuthButtons from '@/src/components/Toolsets/Auth/AuthButtons';
 import { ROOT_FOLDER } from '@/src/constants/file';
 import { useToolsetFolder } from '@/src/context/assets/ToolsetsFolderContext';
