+++ apps/ai-dial-admin/src/components/EntityTabs/Features/constants.ts
@@ -17,6 +17,14 @@ export const placeholdersMap: Record<string, EntityPlaceholdersI18nKey> = {
   configurationEndpoint: EntityPlaceholdersI18nKey.ConfigurationEndpoint,
 };
 
+export const runnerApplicationMap: Record<string, string> = {
+  rateEndpoint: 'dial:applicationTypeRateEndpoint',
+  tokenizeEndpoint: 'dial:applicationTypeTokenizeEndpoint',
+  truncatePromptEndpoint: 'dial:applicationTypeTruncatePromptEndpoint',
+  configurationEndpoint: 'dial:applicationTypeConfigurationEndpoint',
+  assistantAttachmentsInRequestSupported: 'dial:applicationTypeAssistantAttachmentsInRequestSupported',
+};
+
 export const modelsSwitchFeatures: (keyof DialFeatures)[] = [
   'systemPromptSupported',
   'toolsSupported',
