+++ apps/ai-dial-admin/src/components/EntityTabs/Features/utils.ts
@@ -1,7 +1,7 @@
 import { DialApplicationScheme } from '@/src/models/dial/application';
 import { DialFeatures } from '@/src/models/dial/features';
 import { ApplicationRoute } from '@/src/types/routes';
-import { runnerApplicationMap } from './constant';
+import { runnerApplicationMap } from './constants';
 import {
   applicationSwitchFeatures,
   applicationTextFeatures,
