--- apps/ai-dial-admin/src/components/EntityView/Features/constant.ts
@@ -1,7 +1,0 @@
-export const runnerApplicationMap: Record<string, string> = {
-  rateEndpoint: 'dial:applicationTypeRateEndpoint',
-  tokenizeEndpoint: 'dial:applicationTypeTokenizeEndpoint',
-  truncatePromptEndpoint: 'dial:applicationTypeTruncatePromptEndpoint',
-  configurationEndpoint: 'dial:applicationTypeConfigurationEndpoint',
-  assistantAttachmentsInRequestSupported: 'dial:applicationTypeAssistantAttachmentsInRequestSupported',
-};
