+++ apps/ai-dial-admin/src/components/EntityView/View/EntityView.tsx
@@ -6,7 +6,7 @@ import { FC, useCallback, useEffect, useMemo, useState } from 'react';
 import { cloneDeep } from 'lodash';
 
 import HeaderButtons from '@/src/components/EntityView/Header/HeaderButtons';
-import EntityJsonEditor from '@/src/components/EntityView/JsonEditor/JsonEditor';
+import EntityJsonEditor from '@/src/components/EntityTabs/JsonEditor/JsonEditor';
 import { ModalType } from '@/src/components/EntityView/Modals/constants';
 import EntityViewModals from '@/src/components/EntityView/Modals/EntityViewModals';
 import { isDisableRole } from '@/src/components/EntityView/Roles/utils';
