+++ apps/ai-dial-admin/src/components/Routes/View/View.tsx
@@ -8,7 +8,7 @@ import { cloneDeep } from 'lodash';
 import { getCoreRoute, removeRoute, updateCoreRoute, updateRoute } from '@/src/app/[lang]/routes/actions';
 import { JsonConfiguration } from '@/src/components/EntityHeaderControls/models';
 import SimpleEntityHeader from '@/src/components/EntityHeaderControls/SimpleHeader';
-import EntityJsonEditor from '@/src/components/EntityView/JsonEditor/JsonEditor';
+import EntityJsonEditor from '@/src/components/EntityTabs/JsonEditor/JsonEditor';
 import { useNotification } from '@/src/context/NotificationContext';
 import { useProtectedRequest } from '@/src/hooks/use-protected-request';
 import { useI18n } from '@/src/locales/client';
