+++ apps/ai-dial-admin/src/components/Toolsets/View/View.tsx
@@ -8,7 +8,7 @@ import { cloneDeep } from 'lodash';
 import { getCoreToolset, removeToolset, updateCoreToolset, updateToolset } from '@/src/app/[lang]/toolsets/actions';
 import { JsonConfiguration } from '@/src/components/EntityHeaderControls/models';
 import SimpleEntityHeader from '@/src/components/EntityHeaderControls/SimpleHeader';
-import EntityJsonEditor from '@/src/components/EntityView/JsonEditor/JsonEditor';
+import EntityJsonEditor from '@/src/components/EntityTabs/JsonEditor/JsonEditor';
 import EntityRolesModal from '@/src/components/EntityView/Modals/EmptyRoles/EmptyRoles';
 import { isDisableRole } from '@/src/components/EntityView/Roles/utils';
 import { useNotification } from '@/src/context/NotificationContext';
